test: gather npm artifact checks
