+++ test/all.test.js
@@ -22,6 +22,5 @@ import './goods.test.ts'
 import './index.test.js'
 import './log.test.ts'
 import './md.test.ts'
-import './package.test.js'
 import './util.test.js'
 import './vendor.test.js'
